move planet mock out of app into its own component
