+++ frontend/src/App.css
@@ -4,23 +4,8 @@
   padding: 2rem;
   text-align: center;
 }
-.planet{
-  /* width: 200px;
-  height: 200px; */
 
-  border-radius: 50%;
-  border: solid;
-}
-.heavenly-body{
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
 
-.sun {
-  background-color: yellow;
-  color: yellow;
-}
 
 /* .logo {
   height: 6em;
