+++ frontend/src/components/NavBar/NavBar.jsx
@@ -1,0 +1,9 @@
+const NavBar = () => {
+    return (  
+        <>
+        
+        </>
+    );
+}
+ 
+export default NavBar;
