+++ frontend/src/components/Planets/Planets.css
@@ -1,0 +1,21 @@
+.planets {
+	display: flex;
+}
+.planet {
+	/* width: 200px;
+  height: 200px; */
+
+	border-radius: 50%;
+	border: solid;
+}
+.heavenly-body {
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+    justify-content: center;
+}
+
+.sun {
+	background-color: yellow;
+	color: yellow;
+}
