+++ frontend/src/components/Planets/index.js
@@ -1,0 +1,4 @@
+import Planets from "./Planets";
+import  "./Planets.css";
+
+export default Planets;
